Fix pseudocode for compositing step
The pseudocode for compositing was missing one layer allocation. This
would result in incorrectly drawing BackdropPaint just on the previous
state of the canvas, then compositing the source paint down on
everything, instead of compositing backdrop and source first, then
applying the result to the current state of the canvas.